refactor(create-rspeedy): delegate to create-lynx
