+++ packages/rspeedy/create-lynx/rslib.config.ts
@@ -2,8 +2,15 @@ import type { RslibConfig } from '@rslib/core'
 import { defineConfig } from '@rslib/core'
 
 const config: RslibConfig = defineConfig({
+  source: {
+    tsconfigPath: './tsconfig.build.json',
+    entry: {
+      index: './src/index.ts',
+      create: './src/create.ts',
+    },
+  },
   lib: [
-    { format: 'esm', syntax: 'es2021' },
+    { format: 'esm', syntax: 'es2021', dts: true },
   ],
 })
