+++ packages/rspeedy/create-lynx/tsconfig.build.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "types": ["node"],
+  },
+  "include": ["src"],
+}
