--- packages/rspeedy/create-rspeedy/template-common/AGENTS.md
@@ -1,26 +1,0 @@
-# AGENTS.md
-
-You are an expert in JavaScript, Rspeedy, and Lynx application development. You write maintainable, performant, and accessible code.
-
-## Read in Advance
-
-Read docs below in advance to help you understand the library or frameworks this project depends on.
-
-- Lynx: [llms.txt](https://lynxjs.org/next/llms.txt), **REQUIRED**.
-  While dealing with a Lynx task, an agent **MUST** read this doc because it is an entry point of all available docs about Lynx.
-
-## Commands
-
-- `npm run dev` - Start the dev server
-
-- `npm run build` - Build the app for production
-
-- `npm run preview` - Preview the production build locally
-
-- `npm exec rspeedy inspect` - Inspect the Rspeedy config and Rspack config of the project.
-
-## Related Docs
-
-- Rsbuild: <https://rsbuild.rs/llms.txt>
-
-- Rspack: <https://rspack.rs/llms.txt>
