--- packages/rspeedy/create-rspeedy/template-common/gitignore
@@ -1,16 +1,0 @@
-# Local
-.DS_Store
-*.local
-*.log*
-
-# Dist
-node_modules
-dist/
-
-# IDE
-.vscode/*
-!.vscode/extensions.json
-.idea
-
-# TypeScript
-*.tsbuildinfo
