--- packages/rspeedy/create-rspeedy/template-common/src/rspeedy-env.d.ts
@@ -1,14 +1,0 @@
-/// <reference types="@lynx-js/rspeedy/client" />
-
-declare module '@lynx-js/types' {
-  // eslint-disable-next-line @typescript-eslint/no-empty-object-type
-  interface GlobalProps {
-    /**
-     * Define your global properties in this interface.
-     * These types will be accessible through `lynx.__globalProps`.
-     */
-  }
-}
-
-// This export makes the file a module
-export {}
