--- packages/rspeedy/create-rspeedy/template-react-js/lynx.config.js
@@ -1,19 +1,0 @@
-import { defineConfig } from '@lynx-js/rspeedy'
-
-import { pluginQRCode } from '@lynx-js/qrcode-rsbuild-plugin'
-import { pluginReactLynx } from '@lynx-js/react-rsbuild-plugin'
-
-export default defineConfig({
-  source: {
-    entry: './src/index.jsx',
-  },
-  plugins: [
-    pluginQRCode({
-      schema(url) {
-        // We use `?fullscreen=true` to open the page in LynxExplorer in full screen mode
-        return `${url}?fullscreen=true`
-      },
-    }),
-    pluginReactLynx(),
-  ],
-})
