--- packages/rspeedy/create-rspeedy/template-react-js/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "rspeedy-react-js",
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "build": "rspeedy build",
-    "dev": "rspeedy dev",
-    "preview": "rspeedy preview"
-  },
-  "dependencies": {
-    "@lynx-js/react": "workspace:*"
-  },
-  "devDependencies": {
-    "@lynx-js/preact-devtools": "catalog:lynxjs",
-    "@lynx-js/qrcode-rsbuild-plugin": "workspace:*",
-    "@lynx-js/react-rsbuild-plugin": "workspace:*",
-    "@lynx-js/rspeedy": "workspace:*"
-  },
-  "engines": {
-    "node": "^20.19.0 || >=22.12.0"
-  }
-}
