--- packages/rspeedy/create-rspeedy/template-react-js/src/index.jsx
@@ -1,11 +1,0 @@
-import '@lynx-js/preact-devtools'
-import '@lynx-js/react/debug'
-import { root } from '@lynx-js/react'
-
-import { App } from './App.jsx'
-
-root.render(<App />)
-
-if (import.meta.webpackHot) {
-  import.meta.webpackHot.accept()
-}
