--- packages/rspeedy/create-rspeedy/template-react-rstest-rltl/package.json
@@ -1,10 +1,0 @@
-{
-  "scripts": {
-    "test": "rstest run"
-  },
-  "devDependencies": {
-    "@rstest/core": "0.11.12",
-    "@testing-library/jest-dom": "^7.0.0",
-    "jsdom": "^27.4.0"
-  }
-}
