--- packages/rspeedy/create-rspeedy/template-react-rstest-rltl/rstest.config.js
@@ -1,6 +1,0 @@
-import { defineConfig } from '@rstest/core'
-import { withLynxConfig } from '@lynx-js/react/testing-library/rstest-config'
-
-export default defineConfig({
-  extends: withLynxConfig(),
-})
