--- packages/rspeedy/create-rspeedy/template-react-ts/lynx.config.ts
@@ -1,18 +1,0 @@
-import { defineConfig } from '@lynx-js/rspeedy'
-
-import { pluginQRCode } from '@lynx-js/qrcode-rsbuild-plugin'
-import { pluginReactLynx } from '@lynx-js/react-rsbuild-plugin'
-import { pluginTypeCheck } from '@rsbuild/plugin-type-check'
-
-export default defineConfig({
-  plugins: [
-    pluginQRCode({
-      schema(url) {
-        // We use `?fullscreen=true` to open the page in LynxExplorer in full screen mode
-        return `${url}?fullscreen=true`
-      },
-    }),
-    pluginReactLynx(),
-    pluginTypeCheck(),
-  ],
-})
