--- packages/rspeedy/create-rspeedy/template-react-ts/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "rspeedy-react-ts",
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "build": "rspeedy build",
-    "dev": "rspeedy dev",
-    "preview": "rspeedy preview"
-  },
-  "dependencies": {
-    "@lynx-js/react": "workspace:*"
-  },
-  "devDependencies": {
-    "@lynx-js/preact-devtools": "catalog:lynxjs",
-    "@lynx-js/qrcode-rsbuild-plugin": "workspace:*",
-    "@lynx-js/react-rsbuild-plugin": "workspace:*",
-    "@lynx-js/rspeedy": "workspace:*",
-    "@lynx-js/types": "4.1.0",
-    "@rsbuild/plugin-type-check": "1.6.0",
-    "@types/react": "^19.2.18",
-    "typescript": "~6.0.3"
-  },
-  "engines": {
-    "node": "^20.19.0 || >=22.12.0"
-  }
-}
