--- packages/rspeedy/create-rspeedy/template-react-ts/src/index.tsx
@@ -1,11 +1,0 @@
-import '@lynx-js/preact-devtools'
-import '@lynx-js/react/debug'
-import { root } from '@lynx-js/react'
-
-import { App } from './App.js'
-
-root.render(<App />)
-
-if (import.meta.webpackHot) {
-  import.meta.webpackHot.accept()
-}
