--- packages/rspeedy/create-rspeedy/template-react-ts/src/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "composite": true,
-
-    "jsx": "react-jsx",
-    "jsxImportSource": "@lynx-js/react",
-
-    "module": "ESNext",
-    "moduleResolution": "Bundler",
-
-    "noEmit": true,
-  },
-  "include": ["./**/*.ts", "./**/*.tsx"],
-}
