--- packages/rspeedy/create-rspeedy/template-react-ts/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "strict": true,
-    "isolatedModules": true,
-    "verbatimModuleSyntax": true,
-
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-  },
-  "references": [
-    { "path": "./tsconfig.node.json" },
-    { "path": "./src" },
-  ],
-  "files": [],
-}
