--- packages/rspeedy/create-rspeedy/template-react-ts/tsconfig.node.json
@@ -1,18 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "composite": true,
-
-    // Rspeedy uses Node.js's module resolution
-    "module": "node16",
-    "moduleResolution": "node16",
-    "erasableSyntaxOnly": true,
-
-    // Match Rspeedy and Rsbuild v2 runtime requirements.
-    "lib": ["es2023"],
-    "target": "es2022",
-
-    "noEmit": true,
-  },
-  "include": ["./lynx.config.ts"],
-}
