--- packages/rspeedy/create-rspeedy/template-react-vitest-rltl/package.json
@@ -1,10 +1,0 @@
-{
-  "scripts": {
-    "test": "vitest run"
-  },
-  "devDependencies": {
-    "@testing-library/jest-dom": "^7.0.0",
-    "jsdom": "^27.4.0",
-    "vitest": "^3.2.7"
-  }
-}
