--- packages/rspeedy/create-rspeedy/template-react-vitest-rltl/vitest.config.js
@@ -1,9 +1,0 @@
-import { defineConfig } from 'vitest/config'
-import { vitestTestingLibraryPlugin } from '@lynx-js/react/testing-library/plugins'
-
-export default defineConfig({
-  plugins: [
-    vitestTestingLibraryPlugin(),
-  ],
-  test: {},
-})
